move lineup generator to app helper file
